Move steemit_plugin_utiltiies to be a part of steemit_utilities #1496
