+++ libraries/plugins/account_history/CMakeLists.txt
@@ -4,7 +4,7 @@ add_library( account_history_plugin
              account_history_plugin.cpp
            )
 
-target_link_libraries( account_history_plugin chain_plugin steemit_chain steemit_protocol steemit_plugin_utilities )
+target_link_libraries( account_history_plugin chain_plugin steemit_chain steemit_protocol steemit_utilities )
 target_include_directories( account_history_plugin
                             PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include" )
